requirements.txt
streamlit
transformers
torch
Pillow
gtts
deep-translator
